+++ NEWS.md
@@ -9,6 +9,7 @@
 * Computation of `r2()` for `glm`-objects is now based on log-Likelihood methods and also accounts for count models.
 * Better `print()`-method for `overdisp()`.
 * `overdisp()` now supports `glmmTMB`-objects.
+* `boot_ci()` also displays CI based on sample quantiles.
 
 # sjstats 0.10.2